rename quizmachine -> quiz, move messages -> quiz
